Add MerchantCenterAccountInfoCard component.
