--- js/src/components/google-combo-account-card/connect-mc/index.js
@@ -1,1 +1,0 @@
-export { default } from './connect-mc';
