+++ js/src/components/google-mc-account-card/connect-mc/index.js
@@ -1,24 +1,23 @@
 /**
  * External dependencies
  */
-import { CardDivider } from '@wordpress/components';
-import { useState } from '@wordpress/element';
+import { useState, useEffect } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import MerchantCenterSelectControl from '.~/components/merchant-center-select-control';
 import AppButton from '.~/components/app-button';
-import Section from '.~/wcdl/section';
-import Subsection from '.~/wcdl/subsection';
-import ContentButtonLayout from '.~/components/content-button-layout';
-import AccountCard, { APPEARANCE } from '.~/components/account-card';
-import CreateAccountButton from '../create-account-button';
+import AccountCard from '.~/components/account-card';
 import useConnectMCAccount from '.~/hooks/useConnectMCAccount';
-import useCreateMCAccount from '.~/hooks/useCreateMCAccount';
-import AccountConnectionStatus from './account-connection-status';
-import { hasAccountConnectionIssue } from './utils';
+import useGoogleMCAccount from '.~/hooks/useGoogleMCAccount';
+import SwitchUrlCard from '../switch-url-card';
+import ReclaimUrlCard from '../reclaim-url-card';
+import LoadingLabel from '.~/components/loading-label';
+import ConnectedIconLabel from '.~/components/connected-icon-label';
+import MerchantCenterSelect from './merchant-center-select';
+import Actions from './actions';
+import CreatingCard from '../creating-card';
 import './index.scss';
 
 /**
@@ -36,70 +35,133 @@ import './index.scss';
  */
 
 /**
+ * ConnectMC component.
+ *
+ * This component renders Merchant Center connection section.
+ * It is using createMCAccount and resultCreateMCAccount from the parent component.
  * @fires gla_mc_account_connect_button_click
+ * @param {Object} props
+ * @param {Function} props.createMCAccount Callback function for creating a new Merchant Center account.
+ * @param {Object} props.resultCreateMCAccount The result of the create account request.
  */
-const ConnectMC = () => {
+const ConnectMC = ( { createMCAccount, resultCreateMCAccount } ) => {
 	const [ value, setValue ] = useState();
 	const [ handleConnectMC, resultConnectMC ] = useConnectMCAccount( value );
-	const [ handleCreateAccount, resultCreateAccount ] = useCreateMCAccount();
+	const {
+		googleMCAccount,
+		hasFinishedResolution,
+		isReady: isGoogleMCReady,
+	} = useGoogleMCAccount();
 
-	const accountConnectionIssue = hasAccountConnectionIssue(
-		resultConnectMC,
-		resultCreateAccount
-	);
-	if ( accountConnectionIssue ) {
+	useEffect( () => {
+		if ( isGoogleMCReady ) {
+			setValue( googleMCAccount.id );
+		}
+	}, [ googleMCAccount, isGoogleMCReady ] );
+
+	if ( resultConnectMC.response?.status === 409 ) {
 		return (
-			<AccountConnectionStatus
-				resultConnectMC={ resultConnectMC }
-				resultCreateAccount={ resultCreateAccount }
-				onRetry={ handleCreateAccount }
+			<SwitchUrlCard
+				id={ resultConnectMC.error.id }
+				message={ resultConnectMC.error.message }
+				claimedUrl={ resultConnectMC.error.claimed_url }
+				newUrl={ resultConnectMC.error.new_url }
+				onSelectAnotherAccount={ resultConnectMC.reset }
 			/>
 		);
 	}
 
+	if (
+		resultConnectMC.response?.status === 403 ||
+		resultCreateMCAccount.response?.status === 403
+	) {
+		return (
+			<ReclaimUrlCard
+				id={
+					resultConnectMC.error?.id || resultCreateMCAccount.error?.id
+				}
+				websiteUrl={
+					resultConnectMC.error?.website_url ||
+					resultCreateMCAccount.error?.website_url
+				}
+				onSwitchAccount={ () => {
+					resultConnectMC.reset();
+					resultCreateMCAccount.reset();
+				} }
+			/>
+		);
+	}
+
+	if (
+		resultCreateMCAccount.loading ||
+		resultCreateMCAccount.response?.status === 503
+	) {
+		return (
+			<CreatingCard
+				retryAfter={ resultCreateMCAccount.error?.retry_after }
+				onRetry={ createMCAccount }
+			/>
+		);
+	}
+
+	const getIndicator = () => {
+		if ( ! hasFinishedResolution ) {
+			return <LoadingLabel />;
+		}
+
+		if ( isGoogleMCReady ) {
+			return <ConnectedIconLabel />;
+		}
+
+		if ( resultConnectMC.loading ) {
+			return (
+				<LoadingLabel
+					text={ __( 'Connecting…', 'google-listings-and-ads' ) }
+				/>
+			);
+		}
+
+		return (
+			<AppButton
+				isSecondary
+				eventName="gla_mc_account_connect_button_click"
+				eventProps={ { id: Number( value ) } }
+				onClick={ handleConnectMC }
+			>
+				{ __( 'Connect', 'google-listings-and-ads' ) }
+			</AppButton>
+		);
+	};
+
 	return (
 		<AccountCard
 			className="gla-connect-mc-card"
-			appearance={ APPEARANCE.GOOGLE_MERCHANT_CENTER }
-		>
-			<CardDivider />
-			<Section.Card.Body>
-				<Subsection.Title>
-					{ __(
-						'Connect to an existing account',
-						'google-listings-and-ads'
-					) }
-				</Subsection.Title>
-				<ContentButtonLayout>
-					<MerchantCenterSelectControl
-						value={ value }
-						onChange={ setValue }
-					/>
-					<AppButton
-						isSecondary
-						loading={ resultConnectMC.loading }
-						disabled={ ! value }
-						eventName="gla_mc_account_connect_button_click"
-						eventProps={ { id: Number( value ) } }
-						onClick={ handleConnectMC }
-					>
-						{ __( 'Connect', 'google-listings-and-ads' ) }
-					</AppButton>
-				</ContentButtonLayout>
-			</Section.Card.Body>
-			<Section.Card.Footer>
-				<CreateAccountButton
-					isLink
-					disabled={ resultConnectMC.loading }
-					onCreateAccount={ handleCreateAccount }
-				>
-					{ __(
-						'Or, create a new Merchant Center account',
-						'google-listings-and-ads'
-					) }
-				</CreateAccountButton>
-			</Section.Card.Footer>
-		</AccountCard>
+			title={ __(
+				'Connect to existing Merchant Center account',
+				'google-listings-and-ads'
+			) }
+			helper={ __(
+				'Required to sync products so they show on Google.',
+				'google-listings-and-ads'
+			) }
+			alignIndicator="toDetail"
+			indicator={ getIndicator() }
+			detail={
+				<MerchantCenterSelect
+					isConnected={ isGoogleMCReady }
+					value={ value }
+					onChange={ setValue }
+				/>
+			}
+			actions={
+				<Actions
+					isConnected={ isGoogleMCReady }
+					resultConnectMC={ resultConnectMC }
+					resultCreateAccount={ resultCreateMCAccount }
+					handleCreateAccount={ createMCAccount }
+				/>
+			}
+		/>
 	);
 };
