+++ js/src/components/google-mc-account-card/index.js
@@ -1,2 +1,1 @@
-export { default } from './google-mc-account-card';
 export { default as ConnectedGoogleMCAccountCard } from './connected-google-mc-account-card';
