+++ js/src/settings/linked-accounts.js
@@ -19,11 +19,11 @@ import SpinnerCard from '.~/components/spinner-card';
 import VerticalGapLayout from '.~/components/vertical-gap-layout';
 import { ConnectedWPComAccountCard } from '.~/components/wpcom-account-card';
 import { ConnectedGoogleAccountCard } from '.~/components/google-account-card';
-import { ConnectedGoogleMCAccountCard } from '.~/components/google-mc-account-card';
 import { ConnectedGoogleAdsAccountCard } from '.~/components/google-ads-account-card';
 import Section from '.~/wcdl/section';
 import LinkedAccountsSectionWrapper from './linked-accounts-section-wrapper';
 import DisconnectModal, { ALL_ACCOUNTS, ADS_ACCOUNT } from './disconnect-modal';
+import MerchantCenterAccountInfoCard from '.~/components/google-mc-account-card/merchant-center-account-info-card';
 import { GOOGLE_ADS_ACCOUNT_STATUS } from '.~/constants';
 import { queueRecordGlaEvent } from '.~/utils/tracks';
 
@@ -93,9 +93,8 @@ export default function LinkedAccounts() {
 						googleAccount={ google }
 						hideAccountSwitch
 					/>
-					<ConnectedGoogleMCAccountCard
+					<MerchantCenterAccountInfoCard
 						googleMCAccount={ googleMCAccount }
-						hideAccountSwitch
 					/>
 					{ hasAdsAccount && (
 						<ConnectedGoogleAdsAccountCard
